# tour-backend/src/main/resources/openapi/tour-api.yaml
openapi: 3.0.3
info:
  title: Tour Backend API
  version: 0.2.0
  description: API-first backend contract for the tour application.
servers:
  - url: http://localhost:8080/api
    description: Local backend base URL
...
paths:
...
  /editions/current:
    get:
      tags:
        - Public
      operationId: getCurrentEdition
      summary: Get active edition metadata
      responses:
        '200':
          description: Active edition details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditionSummary'
        '404':
          $ref: '#/components/responses/NotFound'
  /editions/{editionId}/riders:
    get:
      tags:
        - Public
      operationId: listRiders
      summary: List riders for an edition
      parameters:
        - $ref: '#/components/parameters/EditionId'
      responses:
        '200':
          description: Riders for the requested edition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /editions/{editionId}/standings:
    get:
      tags:
        - Public
      operationId: getStandings
      summary: Get public standings for an edition
      parameters:
        - $ref: '#/components/parameters/EditionId'
      responses:
        '200':
          description: Standings for the requested edition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandingsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /editions/{editionId}/standings/{playerId}:
    get:
      tags:
        - Public
      operationId: getPlayerStandingDetail
      summary: Get one player standing detail with stage history
      parameters:
        - $ref: '#/components/parameters/EditionId'
        - $ref: '#/components/parameters/PlayerId'
      responses:
        '200':
          description: Player standing detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerStandingDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EditionId:
      name: editionId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    PlayerId:
      name: playerId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
...
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              $ref: '#/components/examples/NotFoundError'
...
  examples:
...
    NotFoundError:
      summary: Resource does not exist
      value:
        code: NOT_FOUND
        message: Requested resource was not found.
        timestamp: '2026-10-01T21:15:00Z'
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: Validation failed
        timestamp:
          type: string
          format: date-time
    EditionSummary:
      type: object
      required:
        - editionId
        - year
        - entryDeadline
        - reserveBudget
        - requiredNationality
        - currentStage
      properties:
        editionId:
          type: integer
          format: int64
          minimum: 1
        year:
          type: integer
          minimum: 2020
        entryDeadline:
          type: string
          format: date-time
        reserveBudget:
          type: integer
          minimum: 0
        requiredNationality:
          type: string
          pattern: '^[A-Z0-9]{3}$'
          description: Three-character nationality code.
        currentStage:
          type: integer
          minimum: 0
        doublePointsStages:
          type: array
          items:
            type: integer
            minimum: 1
    RiderListResponse:
      type: object
      required:
        - editionId
        - riders
      properties:
        editionId:
          type: integer
          format: int64
          minimum: 1
        riders:
          type: array
          items:
            $ref: '#/components/schemas/RiderSummary'
    RiderSummary:
      type: object
      required:
        - riderId
        - gameRiderNumber
        - raceBibNumber
        - name
        - value
        - nationality
      properties:
        riderId:
          type: integer
          format: int64
          minimum: 1
        gameRiderNumber:
          type: integer
          minimum: 1
        raceBibNumber:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
        value:
          type: integer
          minimum: 2
          maximum: 86
        nationality:
          type: string
          pattern: '^[A-Z0-9]{3}$'
        withdrawn:
          type: boolean
          default: false
    StandingsResponse:
      type: object
      required:
        - editionId
        - stage
        - standings
      properties:
        editionId:
          type: integer
          format: int64
          minimum: 1
        stage:
          type: integer
          minimum: 0
        standings:
          type: array
          items:
            $ref: '#/components/schemas/StandingRow'
    StandingRow:
      type: object
      required:
        - rank
        - playerId
        - playerName
        - teamName
        - totalPoints
      properties:
        rank:
          type: integer
          minimum: 1
        playerId:
          type: integer
          format: int64
          minimum: 1
        playerName:
          type: string
        teamName:
          type: string
        totalPoints:
          type: integer
    PlayerStandingDetail:
      type: object
      required:
        - editionId
        - stage
        - player
        - stageHistory
      properties:
        editionId:
          type: integer
          format: int64
          minimum: 1
        stage:
          type: integer
          minimum: 0
        player:
          $ref: '#/components/schemas/StandingRow'
        stageHistory:
          type: array
          items:
            $ref: '#/components/schemas/StageHistoryItem'
    StageHistoryItem:
      type: object
      required:
        - stage
        - stagePoints
        - totalPoints
        - substitutions
        - riderScores
      properties:
        stage:
          type: integer
          minimum: 1
        stagePoints:
          type: integer
        totalPoints:
          type: integer
        substitutions:
          type: array
          items:
            $ref: '#/components/schemas/StageSubstitution'
        riderScores:
          type: array
          items:
            $ref: '#/components/schemas/RiderStageScore'
    StageSubstitution:
      type: object
      required:
        - effectiveStage
        - position
        - replacedRider
        - substituteRider
      properties:
        effectiveStage:
          type: integer
          minimum: 1
        position:
          type: integer
          minimum: 1
          maximum: 11
        replacedRider:
          $ref: '#/components/schemas/StageRiderReference'
        substituteRider:
          $ref: '#/components/schemas/StageRiderReference'
    RiderStageScore:
      type: object
      required:
        - position
        - rider
        - substitute
        - points
        - scoreItems
      properties:
        position:
          type: integer
          minimum: 1
          maximum: 14
        rider:
          $ref: '#/components/schemas/StageRiderReference'
        substitute:
          type: boolean
        points:
          type: integer
        scoreItems:
          type: array
          items:
            $ref: '#/components/schemas/ScoreContribution'
    StageRiderReference:
      type: object
      required:
        - riderId
        - riderName
        - raceBibNumber
      properties:
        riderId:
          type: integer
          format: int64
          minimum: 1
        riderName:
          type: string
        raceBibNumber:
          type: integer
          minimum: 1
    ScoreContribution:
      type: object
      required:
        - ruleCode
        - description
        - points
      properties:
        ruleCode:
          type: string
        description:
          type: string
        points:
          type: integer
...